Update location of some interfaces
