+++ packages/ag-grid-community/src/edit/editApi.ts
@@ -1,9 +1,12 @@
-import type { StartEditingCellParams } from '../api/gridApi';
 import { ensureColumnVisible, ensureIndexVisible } from '../api/scrollApi';
 import type { BeanCollection } from '../context/context';
 import { _getRowNode } from '../entities/positionUtils';
 import type { RowNode } from '../entities/rowNode';
-import type { EditingCellPosition, ICellEditorValidationError } from '../interfaces/iCellEditor';
+import type {
+    EditingCellPosition,
+    ICellEditorValidationError,
+    StartEditingCellParams,
+} from '../interfaces/iCellEditor';
 import type { CellPosition } from '../interfaces/iCellPosition';
 import type { IRowNode } from '../interfaces/iRowNode';
 import { _warn } from '../validation/logging';
