+++ packages/ag-grid-community/src/entities/rowNode.ts
@@ -1,6 +1,5 @@
 import { LocalEventService } from '../agStack/events/localEventService';
 import type { IAgEventEmitter, IEventEmitter } from '../agStack/interfaces/iEventEmitter';
-import type { DetailGridInfo } from '../api/gridApi';
 import type { BeanCollection } from '../context/context';
 import type { SelectionEventSourceType } from '../events';
 import { _getRowIdCallback } from '../gridOptionsUtils';
@@ -15,6 +14,7 @@ import type {
     RowNodeEventType,
     RowPinnedType,
 } from '../interfaces/iRowNode';
+import type { DetailGridInfo } from '../interfaces/masterDetail';
 import { _error, _warn } from '../validation/logging';
 import type { AgColumn } from './agColumn';
