+++ packages/ag-grid-community/src/misc/apiEvents/apiEventService.ts
@@ -1,8 +1,8 @@
 import type { NamedBean } from '../../context/bean';
 import { BeanStub } from '../../context/beanStub';
 import type { AgEventType } from '../../eventTypes';
+import { ALWAYS_SYNC_GLOBAL_EVENTS } from '../../eventTypes';
 import type { AgEventListener, AgGlobalEventListener } from '../../events';
-import { ALWAYS_SYNC_GLOBAL_EVENTS } from '../../events';
 import type { IFrameworkEventListenerService } from '../../interfaces/iFrameworkEventListenerService';
 
 export class ApiEventService extends BeanStub<AgEventType> implements NamedBean {
